GPGPU variant of LCG32.
Add GPGPU support using HIP.
Extend cmake configuration files.

+++ utility/simd/simd.hpp
@@ -77,7 +77,7 @@ namespace XXX_NAMESPACE
         template <>                                                 \
         struct Type<T>                                              \
         {                                                           \
-            static constexpr std::size_t width = WIDTH;             \
+            static constexpr std::int32_t width = WIDTH;            \
         };
 
         MACRO(double, SIMD_WIDTH_NATIVE_64BIT)
